wheelwhere_server/requirements.txt
amqp==2.5.2
beautifulsoup4==4.8.2
billiard==3.6.3.0
boto3==1.14.13
botocore==1.17.13
bs4==0.0.1
celery==4.4.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.5
django-aligo==0.1.2
django-appconf==1.0.3
django-cors-headers==3.4.0
django-debug-toolbar==2.2
django-extensions==2.2.9
django-imagekit==4.0.2
django-import-export==2.0.2
django-mail-admin==0.1.4
django-rest-framework==0.1.0
djangorestframework==3.11.0
docutils==0.15.2
et-xmlfile==1.0.1
idna==2.9
importlib-metadata==1.5.2
jdcal==1.4.1
jmespath==0.10.0
jwt==0.6.1
kombu==4.6.8
MarkupPy==1.14
mysqlclient==1.4.6
odfpy==1.4.1
openpyxl==3.0.3
pilkit==2.0
Pillow==7.0.0
pycparser==2.20
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tablib==1.1.0
urllib3==1.25.8
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
